requirements.txt
streamlit
tensorflow
opencv-python
pillow
numpy
streamlit-webrtc
